requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.1
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
pandas==2.0.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.17
stripe==5.5.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
